Extract common metrics functionality to a new module
